Move product to own project.
This is needed because of the otherwise circular referencing:

the application feature references the application within which the
product file is contained and the product file references the
application feature.

By moving the product file to a separate project we break this cycle.

OPEN - task 2: Add possibility to build and export with maven 
https://github.com/mudelta/e4base/issues/issue/2

Signed-off-by: Davy Meers <DFJ.Meers@mindef.nl>