Add example to investigate possible Tycho problem
Building net.davymeers.e4.base.product fails.
In the logging are warnings from the mirror tool about "Problems
resolving provisioning plan". 
The added link[1] explains that in case the warning only reports
unsatisfied dependencies to the original versions of plug-ins replaced
through a feature patch it can be safely ignored.

I am not sure it this is the case. 
This commit is meant to show the problem so that it can be investigated.

[1]http://wiki.eclipse.org/Tycho_Messages_Explained#Mirror_tool



Signed-off-by: Davy Meers <DFJ.Meers@mindef.nl>