Fix runtime configuration + Mavenize Product Project
CLOSED task 2: Add possibility to build and export with maven 
https://github.com/mudelta/e4base/issues/issue/2

In order to run the e4 application a declarative service implementation
is needed.
The org.eclipse.equinox.ds bundle is added to the application feature
and auto-started.
Also, the eclipse runtime (org.eclipse.core.runtime) has to be
auto-started.

The Product project is mavenized and a POM is added.
Furthermore the product module is added to the aggregator POM.

This fixes #2.

Signed-off-by: Davy Meers <DFJ.Meers@mindef.nl>